Read IA32_EFER #9
